Mass rename of files from *.plist to *.tmSnippet
I have no idea why there are so many .plist files in the Snippets directory when TextMate creates .tmSnippet. So, in order to tidy up the repo and to make it easier to edit the files in a text editor rather than the Property List Editor I thought it best to do a rename now rather than later.